feature: Added loader to tv show page
