+++ src/App.vue
@@ -3,5 +3,3 @@
 <template>
   <RouterView />
 </template>
-
-<style scoped></style>
